Move k8s client to pkg
